+++ src/meltano/api/controllers/dashboards.py
@@ -1,8 +1,10 @@
 from flask import jsonify, request
-from .dashboards_helper import (
+from .dashboards_helper import DashboardReportScheduleNotFoundError, DashboardsHelper
+from meltano.core.project import Project
+from meltano.core.m5o.dashboards_service import (
     DashboardAlreadyExistsError,
     DashboardDoesNotExistError,
-    DashboardsHelper,
+    DashboardsService,
 )
 from .errors import InvalidFileNameError
 from meltano.api.api_blueprint import APIBlueprint
@@ -12,9 +14,14 @@ from meltano.api.security.readonly_killswitch import readonly_killswitch
 dashboardsBP = APIBlueprint("dashboards", __name__)
 
 
+def dashboards_service():
+    project = Project.find()
+    return DashboardsService(project)
+
+
 @dashboardsBP.errorhandler(DashboardAlreadyExistsError)
 def _handle(ex):
-    dashboard_name = ex.dashboard_name
+    dashboard_name = ex.dashboard["name"]
     return (
         jsonify(
             {
@@ -50,17 +57,28 @@ def _handle(ex):
     )
 
 
+@dashboardsBP.errorhandler(DashboardReportScheduleNotFoundError)
+def _handle(ex):
+    return (
+        jsonify(
+            {
+                "error": True,
+                "code": f"The reports could not be loaded because a pipeline for namespace '{ex.namespace}' could not be found.",
+            }
+        ),
+        404,
+    )
+
+
 @dashboardsBP.route("/all", methods=["GET"])
 def get_dashboards():
-    dashboards_helper = DashboardsHelper()
-    response_data = dashboards_helper.get_dashboards()
+    response_data = dashboards_service().get_dashboards()
     return jsonify(response_data)
 
 
 @dashboardsBP.route("/dashboard/<dashboard_id>", methods=["GET"])
 def get_dashboard(dashboard_id):
-    dashboards_helper = DashboardsHelper()
-    response_data = dashboards_helper.get_dashboard(dashboard_id)
+    response_data = dashboards_service().get_dashboard(dashboard_id)
     return jsonify(response_data)
 
 
@@ -70,9 +88,8 @@ def save_dashboard():
     """
     Endpoint for saving a dashboard
     """
-    dashboards_helper = DashboardsHelper()
     post_data = request.get_json()
-    response_data = dashboards_helper.save_dashboard(post_data)
+    response_data = dashboards_service().save_dashboard(post_data)
     return jsonify(response_data)
 
 
@@ -82,9 +99,8 @@ def delete_dashboard():
     """
     Endpoint for deleting a dashboard
     """
-    dashboards_helper = DashboardsHelper()
     post_data = request.get_json()
-    response_data = dashboards_helper.delete_dashboard(post_data)
+    response_data = dashboards_service().delete_dashboard(post_data)
     return jsonify(response_data)
 
 
@@ -94,27 +110,24 @@ def update_dashboard():
     """
     Endpoint for updating a dashboard
     """
-    dashboards_helper = DashboardsHelper()
     post_data = request.get_json()
-    response_data = dashboards_helper.update_dashboard(post_data)
+    response_data = dashboards_service().update_dashboard(post_data)
     return jsonify(response_data)
 
 
 @dashboardsBP.route("/dashboard/report/add", methods=["POST"])
 @readonly_killswitch
 def add_report_to_dashboard():
-    dashboards_helper = DashboardsHelper()
     post_data = request.get_json()
-    response_data = dashboards_helper.add_report_to_dashboard(post_data)
+    response_data = dashboards_service().add_report_to_dashboard(post_data)
     return jsonify(response_data)
 
 
 @dashboardsBP.route("/dashboard/report/remove", methods=["POST"])
 @readonly_killswitch
 def remove_report_from_dashboard():
-    dashboards_helper = DashboardsHelper()
     post_data = request.get_json()
-    response_data = dashboards_helper.remove_report_from_dashboard(post_data)
+    response_data = dashboards_service().remove_report_from_dashboard(post_data)
     return jsonify(response_data)