+++ src/meltano/api/controllers/repos.py
@@ -7,7 +7,6 @@ import markdown
 from flask import jsonify, request
 
 from meltano.core.project import Project
-from meltano.core.utils import decode_file_path_from_id
 from meltano.core.compiler.project_compiler import ProjectCompiler
 from meltano.api.api_blueprint import APIBlueprint
 from meltano.core.m5o.m5o_file_parser import (
@@ -81,7 +80,9 @@ def index():
     onlydocs = project.model_dir().parent.glob("*.md")
 
     for d in onlydocs:
-        file_dict = MeltanoAnalysisFileParser.fill_base_m5o_dict(d, str(d.name))
+        file_dict = MeltanoAnalysisFileParser.fill_base_m5o_dict(
+            d.relative_to(project.root), str(d.name)
+        )
         sortedM5oFiles["documents"]["items"].append(file_dict)
 
     for f in onlyfiles:
@@ -92,7 +93,9 @@ def index():
         # filename splittext twice occurs due to current *.type.extension convention (two dots)
         filename = filename.lower()
         filename, ext = os.path.splitext(filename)
-        file_dict = MeltanoAnalysisFileParser.fill_base_m5o_dict(f, filename)
+        file_dict = MeltanoAnalysisFileParser.fill_base_m5o_dict(
+            f.relative_to(project.root), filename
+        )
         if ext == ".topic":
             sortedM5oFiles["topics"]["items"].append(file_dict)
         if ext == ".table":
@@ -115,28 +118,6 @@ def index():
         )
 
     return jsonify(sortedM5oFiles)
-
-
-@reposBP.route("/file/<unique_id>", methods=["GET"])
-def file(unique_id):
-    file_path = decode_file_path_from_id(unique_id)
-    (filename, ext) = os.path.splitext(file_path)
-    is_markdown = False
-    project = Project.find()
-    path_to_file = project.model_dir(file_path).resolve()
-    with open(path_to_file, "r") as read_file:
-        data = read_file.read()
-        if ext == ".md":
-            data = markdown.markdown(data)
-            is_markdown = True
-        return jsonify(
-            {
-                "file": data,
-                "is_markdown": is_markdown,
-                "id": unique_id,
-                "populated": True,
-            }
-        )
 
 
 def lint_all(compile):