+++ src/webapp/src/api/repo.js
@@ -2,10 +2,6 @@ import axios from 'axios'
 import utils from '@/utils/utils'
 
 export default {
-  file(id) {
-    return axios.get(utils.apiUrl('repos', `file/${id}`))
-  },
-
   index() {
     return axios.get(utils.apiUrl('repos'))
   },